requirements.txt
attrs==19.3.0
Automat==20.2.0
cffi==1.14.0
click==7.1.1
constantly==15.1.0
cryptography==2.9.1
cssselect==1.1.0
Flask==1.1.2
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.0
MarkupSafe==1.1.1
parsel==1.5.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-decouple==3.3
queuelib==1.5.0
Scrapy==2.0.1
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
Twisted==20.3.0
urllib3==1.25.9
w3lib==1.21.0
Werkzeug==1.0.1
zope.interface==5.1.0
